4-tier-todo-app/app/crud/requirements.txt
Flask
mysql-connector-python
retry